backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
requests
pandas
Faker
sentence-transformers
faiss-cpu
numpy
